# OneDrive/Área de Trabalho/Portifolio1/swagger.yaml
openapi: 3.0.3
info:
  title: To-Do List API
  description: API para gerenciamento de tarefas pessoais com operações de criação, edição, listagem, filtro por status e persistência.
  version: 1.0.0
servers:
  - url: http://localhost:4173
    description: Servidor local da aplicação
paths:
  /api/tasks:
    get:
      summary: Listar tarefas
      description: Retorna todas as tarefas ou apenas as tarefas por status.
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [all, active, completed]
            default: all
          description: Filtra a lista por status da tarefa.
      responses:
        '200':
          description: Lista de tarefas retornada com sucesso.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Criar uma nova tarefa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Tarefa criada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/tasks/{id}:
    get:
      summary: Buscar tarefa por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Identificador da tarefa.
      responses:
        '200':
          description: Tarefa retornada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Atualizar tarefa
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Tarefa atualizada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Excluir tarefa
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Tarefa excluída com sucesso.
        '404':
          $ref: '#/components/responses/NotFound'

  /api/tasks/{id}/toggle:
    patch:
      summary: Alternar status da tarefa
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Status da tarefa alterado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          example: "t_123456"
        title:
          type: string
          example: "Estudar QA"
        completed:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
          example: "2026-08-13T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-08-13T10:35:00Z"
      required:
        - id
        - title
        - completed
        - createdAt
        - updatedAt

    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
          example: "Fazer revisão do código"
        completed:
          type: boolean
          example: false
      required:
        - title

    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
          example: "Atualizar checklist da sprint"
        completed:
          type: boolean
          example: true

    Error:
      type: object
      properties:
        message:
          type: string
          example: "Tarefa não encontrada."
      required:
        - message

  responses:
    BadRequest:
      description: Requisição inválida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso não encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'